+++ yarn-project/p2p/src/test-helpers/index.ts
@@ -4,3 +4,4 @@ export * from './make-enrs.js';
 export * from './make-test-p2p-clients.js';
 export * from './reqresp-nodes.js';
 export * from './mock-pubsub.js';
+export * from './testbench-utils.js';
